todoapp-master/requirements.txt
asgiref==3.3.1
Django==3.1.4
django-autoslug==1.9.8
django-crispy-forms==1.10.0
djangorestframework==3.12.2
python-memcached==1.59
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
